fix(backend): repair db container initialization
When the db container was initialized, the backend container was not waiting for the db container to be ready. This caused the backend container to fail to start. The wait-for-it.sh script was added to the backend container to wait for the db container to be ready before starting the backend container.